requirements.txt
absl-py==1.0.0
aenum==3.1.5
anyio==3.4.0
appdirs==1.4.4
argon2-cffi @ file:///tmp/build/80754af9/argon2-cffi_1613037499734/work
asdf==2.8.2
astropy @ file:///tmp/build/80754af9/astropy_1629820331715/work
astunparse==1.6.3
async-generator @ file:///home/ktietz/src/ci/async_generator_1611927993394/work
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
awkward==1.4.0
awkward0==0.15.5
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
bidict==0.21.4
bleach @ file:///tmp/build/80754af9/bleach_1628110601003/work
boost-histogram==1.1.0
cachetools==4.2.4
certifi==2021.10.8
cffi @ file:///tmp/build/80754af9/cffi_1625814692085/work
charset-normalizer==2.0.8
click==8.0.3
click-default-group==1.2.2
cloup==0.7.1
colorama==0.4.4
colorlog==6.6.0
colorspacious==1.1.2
colour==0.1.5
commonmark==0.9.1
cupy==9.6.0
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
Cython==0.29.24
debugpy @ file:///tmp/build/80754af9/debugpy_1637091799509/work
DecayLanguage==0.12.0
decorator @ file:///tmp/build/80754af9/decorator_1632776554403/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
Deprecated==1.2.13
docutils==0.18.1
entrypoints==0.3
et-xmlfile==1.1.0
fastrlock==0.8
ffmpeg-python==0.2.0
flatbuffers==2.0
fonttools==4.25.0
formulate==0.1.1
future==0.18.2
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.24
glcontext==2.3.4
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.19.1
grpcio==1.43.0
h5py==3.6.0
healpy==1.15.0
hepstats==0.5.0
hepunits==2.1.2
hist==2.4.0
histoprint==2.2.0
idna==3.3
iminuit==2.8.4
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1631916692253/work
ipykernel @ file:///tmp/build/80754af9/ipykernel_1633534655931/work/dist/ipykernel-6.4.1-py3-none-any.whl
ipython @ file:///tmp/build/80754af9/ipython_1635944169458/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1634143127070/work
isosurfaces==0.1.0
jedi @ file:///tmp/build/80754af9/jedi_1611333729159/work
Jinja2 @ file:///tmp/build/80754af9/jinja2_1635780242639/work
jmespath==0.10.0
jsonschema @ file:///Users/ktietz/demo/mc3/conda-bld/jsonschema_1630511932244/work
jupyter @ file:///tmp/build/80754af9/jupyter_1607700846274/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1637148478538/work
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1616615302928/work
jupyter-core @ file:///tmp/build/80754af9/jupyter_core_1636537028672/work
jupyter-manim==1.3
jupyter-server==1.13.1
jupyter-server-mathjax==0.2.3
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-widgets @ file:///tmp/build/80754af9/jupyterlab_widgets_1609884341231/work
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1612281846643/work
kneed==0.7.0
lark-parser==0.8.5
libclang==12.0.0
llvmlite==0.36.0
manim==0.13.1
manimlib==0.2.0
ManimPango==0.3.1
mapbox-earcut==0.12.11
Markdown==3.3.6
MarkupSafe @ file:///tmp/build/80754af9/markupsafe_1621523467000/work
matplotlib==3.4.3
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mistune @ file:///tmp/build/80754af9/mistune_1607364877025/work
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
moderngl==5.6.4
moderngl-window==2.4.1
mplhep==0.3.15
mplhep-data==0.0.3
mpmath==1.2.1
multipledispatch==0.6.0
munkres==1.1.4
nb-conda-kernels @ file:///tmp/build/80754af9/nb_conda_kernels_1636035787744/work
nbclient @ file:///tmp/build/80754af9/nbclient_1614364831625/work
nbconvert @ file:///tmp/build/80754af9/nbconvert_1624472883256/work
nbdime==3.1.1
nbformat @ file:///tmp/build/80754af9/nbformat_1617383369282/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
networkx @ file:///tmp/build/80754af9/networkx_1633639043937/work
notebook @ file:///tmp/build/80754af9/notebook_1637161762562/work
numba==0.53.0
numpy==1.21.5
oauthlib==3.1.1
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
opencv-python==4.5.4.60
openpyxl==3.0.9
opt-einsum==3.3.0
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.3.4
pandocfilters @ file:///tmp/build/80754af9/pandocfilters_1605120906940/work
parso @ file:///tmp/build/80754af9/parso_1617223946239/work
particle==0.16.2
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.4.0
plumbum==1.7.1
pooch==1.5.2
progressbar==2.5
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1637050397234/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.19.1
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAstronomy==0.17.0
pybind11==2.8.1
pycairo==1.20.1
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydub==0.25.1
pyerfa @ file:///tmp/build/80754af9/pyerfa_1621556109336/work
pyglet==1.5.21
Pygments @ file:///tmp/build/80754af9/pygments_1629234116488/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
PyPDT==0.8.0
pyrr==0.10.3
pyrsistent @ file:///tmp/build/80754af9/pyrsistent_1636110951836/work
pyshtools==4.8
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
PyYAML==6.0
pyzmq @ file:///tmp/build/80754af9/pyzmq_1628275385016/work
qtconsole @ file:///tmp/build/80754af9/qtconsole_1632739723211/work
QtPy @ file:///tmp/build/80754af9/qtpy_1629397026935/work
quantities==0.12.5
qutip==4.6.2
requests==2.26.0
requests-oauthlib==1.3.0
rich==10.16.1
rsa==4.8
scikit-hep==3.4.0
scipy @ file:///tmp/build/80754af9/scipy_1630606796912/work
screeninfo==0.6.7
semantic-version==2.8.5
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1623709665295/work
skia-pathops==0.7.2
smmap==5.0.0
sniffio==1.2.0
srt==3.5.0
statistics==1.0.3.5
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
terminado==0.9.4
testpath @ file:///tmp/build/80754af9/testpath_1624638946665/work
tornado @ file:///tmp/build/80754af9/tornado_1606942317143/work
tqdm==4.62.3
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
typing_extensions==4.0.1
uhi==0.3.0
uproot==4.0.11
uproot3==3.14.4
uproot3-methods==0.10.1
urllib3==1.26.7
vector==0.8.4
watchdog==2.1.6
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.2
widgetsnbextension @ file:///tmp/build/80754af9/widgetsnbextension_1607531506226/work
wrapt==1.13.3
xarray==0.20.1
xlwt==1.3.0
zipp @ file:///tmp/build/80754af9/zipp_1633618647012/work
